send only if open and change in status
