+++ config/vars.template.json
@@ -1,7 +1,7 @@
 {
   "TARGET_GUILD_ID": "GUILD_ID",
   "COFFEE_ROLE_ID": "ROLE_ID",
-  "OFFICE_PING_ROLE_ID": "PING_ID",
+  "OFFICE_PING_ROLE_ID": "ROLE_ID",
   "NOTIF_CHANNEL_ID": "CHANNEL_ID",
   "ANNOUNCEMENTS_CHANNEL_ID": "CHANNEL_ID",
   "OFFICE_STATUS_CHANNEL_ID": "CHANNEL_ID",
