+++ src/components/cron.ts
@@ -2,8 +2,8 @@ import { CronJob } from 'cron';
 import { Client, MessageEmbed, TextChannel } from 'discord.js';
 import _ from 'lodash';
 import fetch from 'node-fetch';
-import { alertMatches } from '../commands/coffeeChat/coffeeChat';
-import { alertUsers } from '../commands/officeOpenDM/officeOpenDM';
+import { alertMatches } from '../components/coffeeChat';
+import { alertUsers } from './officeOpenDM';
 import { vars } from '../config';
 import { DEFAULT_EMBED_COLOUR } from '../utils/embeds';
 import { getMatch, writeHistoricMatches } from '../components/coffeeChat';
@@ -19,6 +19,11 @@ import {
 const NOTIF_CHANNEL_ID: string = vars.NOTIF_CHANNEL_ID;
 const OFFICE_STATUS_CHANNEL_ID: string = vars.OFFICE_STATUS_CHANNEL_ID;
 const OFFICE_HOURS_STATUS_API = 'https://csclub.ca/office-status/json';
+
+// The last known status of the office
+//  false if closed
+//  true if open
+let office_last_status = false;
 
 export const initCrons = async (client: Client): Promise<void> => {
   createSuggestionCron(client).start();
@@ -53,10 +58,15 @@ export const createOfficeStatusCron = (client: Client): CronJob =>
       throw 'Bad channel type';
     }
 
-    if (response['status'] == 1) {
+    if (office_last_status == false && response['status'] == 1) {
+      // The office was closed and is now open
       // Send all the users with the "Office Ping" role a DM:
       // Get all users with "Office Ping" role
       await alertUsers();
+      office_last_status = true;
+    } else if (office_last_status == true && response['status'] == 0) {
+      // the office was open and is now closed
+      office_last_status = false;
     }
   });
